src/main/java/com/sasha/course/dao/StudentDAOImpl.java
package com.sasha.course.dao;

import com.sasha.course.entities.Student.Student;
import jakarta.persistence.EntityManager;
import jakarta.persistence.Query;
import jakarta.persistence.TypedQuery;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Repository
public class StudentDAOImpl implements StudentDAO {
    private final EntityManager entityManager;

    @Autowired
    public StudentDAOImpl(EntityManager entityManager) {
        this.entityManager = entityManager;
    }

    @Override
    public void save(Student student) {
        entityManager.persist(student);
    }

    @Override
    public Student findById(int id) {
        return entityManager.find(Student.class, id);
    }

    @Override
    public List<Student> findByLastName(String lastName) {
        TypedQuery<Student> query = entityManager.createQuery("FROM Student WHERE lastName = :lastName order by firstName", Student.class);
        query.setParameter("lastName", lastName);
        return query.getResultList();
    }

    @Override
    public List<Student> findAll() {
        TypedQuery<Student> query = entityManager.createQuery("FROM Student", Student.class);
        return query.getResultList();
    }

    @Override
    @Transactional
    public Student update(Student student) {
        return entityManager.merge(student);
    }

    @Override
    @Transactional
    public int updateAllLastNames(String oldLastName, String newLastName) {
        Query query = entityManager.createQuery("UPDATE Student SET lastName = :newLastName WHERE lastName = :oldLastName");
        query.setParameter("newLastName", newLastName);
        query.setParameter("oldLastName", oldLastName);
        return query.executeUpdate();
    }

    @Override
    @Transactional
    public void delete(Student student) {
        entityManager.remove(student);
    }
}
